+++ R/location_sig.R
@@ -1,4 +1,13 @@
-location_sig <- function(X,y,priors,BTE = c(3000,100000,1), verb =1){
+location_sig <- function(X,y,priors,BTE = c(3000,100000,1), verb =1, eps = sqrt(.Machine$double.eps)){
+
+  nDigits <- function(x){
+    truncX <- floor(abs(x))
+    if(truncX != 0){
+      floor(log10(truncX)) + 1
+    } else {
+      1
+    }
+  }
 
   ## Tests
   K <- ncol(y[[1]])
@@ -56,22 +65,38 @@ location_sig <- function(X,y,priors,BTE = c(3000,100000,1), verb =1){
 
   beta <- matrix(NA, nrow = no.betas*M, ncol = iters)
 
-  bhat <- as.matrix(solve(t(X) %*% X) %*% t(X) %*% Y)
-  bhat[which(bhat < 0)] <- 1000
+  bhat <- as.vector(solve(t(X) %*% X) %*% t(X) %*% Y)
+  bhat[which(bhat < 0)] <- 0
 
-  beta[,1] <- bhat
 
-  if(is.na(priors)){
-  S.prior <- lapply(y.reorg,function(X){diag((apply(X,1,sd))^2)})
-  nu0 <- M
-  mu0 <- rep(0, times = no.betas*M)
-  V0i <- diag(no.betas*M)/10000000
+  if(priors == "Jeffreys"){
+    mu0 <- rep(0, length(bhat))
+    dgts <- sapply(bhat,nDigits)
+    V0i <- diag(length(dgts))*0
+    nu0 <- 0
+    S.prior <- replicate(N, matrix(0, nrow = M, ncol = M), simplify = FALSE)
+    if(verb != 0){message("Jeffreys Priors Used\n")}
+  }else if(is.list(priors)){
+      nu0 <- priors$Sig$nu0
+      mu0 <- priors$beta$mu0
+      V0i <- solve(priors$beta$V0)
+      S.prior <- priors$Sig$S
+      if(verb != 0){message("User Specified Priors Used\n")}
   }else{
-    nu0 <- priors$Sig$nu0
-    mu0 <- priors$beta$mu0
-    V0i <- solve(priors$beta$V0)
-    S.prior <- priors$Sig$S
-    }
+    S.prior <- lapply(y.reorg,function(X){diag((apply(X,1,sd))^2) + diag(nrow(X))*eps})
+    nu0 <- M
+    mu0 <- bhat
+    dgts <- sapply(bhat,nDigits)
+    V0i <- diag(1/(10^(dgts+6)))
+    if(verb != 0){message("Default Priors Used\n")}
+  }
+
+  for(i in 1:N){
+    S.prior[[i]][S.prior[[i]] == eps] <- 1/nu0
+  }
+
+  bhat[which(bhat == 0)] <- eps
+  beta[,1] <- bhat
 
   V0imu0 <- V0i %*% mu0
 
@@ -130,5 +155,5 @@ location_sig <- function(X,y,priors,BTE = c(3000,100000,1), verb =1){
 
   return(list(beta = beta.return,
               Sig = Sig,
-              priors= list(beta = list(mu0 = mu0, V0 = solve(V0i)), Sig = list(nu0 = nu0, S = S.prior)), cov.structure = "location", y.cov = cov.names))
+              priors= list(beta = list(mu0 = mu0, V0 = diag(1/diag(V0i))), Sig = list(nu0 = nu0, S = S.prior)), cov.structure = "location", y.cov = cov.names))
 }